+++ app/models/prep_kitchen/sales_point_link.rb
@@ -9,6 +9,7 @@ module PrepKitchen
 
     validates :sales_point_tenant_id, uniqueness: true
     validate :tenant_types
+    validate :same_organization
 
     private
 
@@ -18,6 +19,19 @@ module PrepKitchen
       end
       unless sales_point_tenant&.sales_point?
         errors.add(:sales_point_tenant, "must be sales_point")
+      end
+    end
+
+    def same_organization
+      return if prep_kitchen_tenant.blank? || sales_point_tenant.blank?
+
+      kitchen_org_id = prep_kitchen_tenant.organization_id
+      point_org_id = sales_point_tenant.organization_id
+
+      if kitchen_org_id.blank?
+        errors.add(:prep_kitchen_tenant, "must belong to an organization")
+      elsif point_org_id.blank? || point_org_id != kitchen_org_id
+        errors.add(:sales_point_tenant, "must belong to the same organization as prep kitchen")
       end
     end
   end